Finalisation liam (#163)
* pricing page / email updates

Added to pricing page - will be linked from marketing site; closes #140
Added radio buttons in this branch on creating job offers;
Fixed e-mail for employers (welcome);
Changed view accepted offers button on employer page;

* Restored data preservation on joboffers

* secondary job field properly returns null unselected

* Secured controllers

Also restored admin functionality of deleting comments

* Rendancy removed

+++ app/Http/Controllers/EmployerCommentsController.php
@@ -88,9 +88,14 @@ class EmployerCommentsController extends Controller
 
     public function destroy($id)
     {
+        
         $comment = Comment::find($id);
+        if(Auth::user()->employer->id == $comment->employer_id){
         $comment->delete();
         return redirect('/reviews');
+        }
+        else
+        return view('home');
     }
